requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-jwt-extended==4.3.1
psycopg2-binary>=2.9.5
requests==2.26.0
gunicorn==20.1.0
python-dotenv==0.19.1
Werkzeug==2.0.3
flask-migrate==3.1.0
flask-cors==3.0.10
SQLAlchemy==1.4.22
